Pipeline syntax generator visualization
I added support for the Pipeline Syntax snippet generator.
This basically means that there are some Jelly and HTML files in the resources
folder that help Jenkins figure out what the possible parameters are help the
users figure out which ones they need to use and how.

I only supported `withAWS`, `s3Download`, and `s3Upload`, since those are the
only ones that I have experience with.  Doing the others should be easy, but
I'll leave that for another time.

Finally, I noticed that if the S3 upload path begins with a "/", then a "//"
will be written into the final path string, and S3 will actually create a zero-length
folder name, which is definitely not what people want.  So, I updated README.md and
the help text for the `path` parameter accordingly.