Added new step: s3Delete
This deletes the object stored in the bucket at the given path.
If the "recursive" flag is set, then this interprets the path as a folder
and deletes all of the objects stored "within" that folder.

Of course, S3 does not actually contain folders; it is a flat structure.
However, S3 has full support for using it to contain pseudo-hierarchies
using a delimiter and some "prefix" logic.  Since the "s3Upload" and "s3Download"
steps already support this with the "/" delimiter, "s3Delete" does so as well.

In terms of logging, "s3Delete" will log analogously to "s3Upload" and "s3Download";
that is, it will log the beginning and end of the operation.  In addition, it will
log when it *actually* deletes an object.