Workaround for the fake folder stuff.
Basically, AWS lets you create anything with any name; it's a key/value filestore.
However, the convention (including in the S3 Management Console) is that "/" is a
name delimiter for paths.  Anywhere that a key ends in "/", it is considered a
"folder", even if it is only a virtual one (through the "prefix" logic).

Now, if you create a folder through the console, it creates a zero-length file whose
name ends with "/".  Thus, the console sees it as an empty folder, even though it
is really a zero-length file.

Here, I excluded those objects entirely.  They will never be returned as "files" from
`s3FindFiles`.  They will, however, continue to show up as folders.