Manually render the paths using "/" as the separator.
This wasn't an issue on Linux machines, but it was a problem on Windows.  Because we were
using the `Path` class to handle the path-related logic, and because the `Path` instances
are created using the "default" `FileSystem` instance, the separator between path parts
varies based on OS.

So, what we do now is just manually replace all instances of `File.separator` with "/"
when we create the `FileWrapper` instances.  This should remove any ambiguity, and it
also allows for an easy fix in the future if AWS ends up changing the separator.