Fix #210 with changesets
Change sets have a status that needs to be validated on whether the
change set was successful or not.

1) Failed, with a message that the changeset was empty
2) success, with no changes (such as output changes)

Previously, the status was not consulted, only the number of changes.
This did not correctly apply changes when it is only the output that
changes.

The previous attempt to fix this issue did not correctly handle an empty
changeset that failed to create due to no changes.